Refaktoroi lops2021 osasuoritusvalidaatiot omaan objektiinsa;
Muuta luokka objektiksi, koska vaikuttaa olevan käytetympi pattern
koodissamme. Poistaa samalla turhan rekursiivisesti usein kutsuttavan
konstruktorikutsun.
